pyproject.toml
[project]
name = "hackerrank-orchestrate-may26"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.40.0",
    "chromadb>=0.5.23",
    "openai>=1.50.0",
    "pandas>=2.2.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "rich>=13.9.0",
    "sentence-transformers>=3.3.0",
    "tqdm>=4.67.0",
]
